Move TestingSessionContext to presto-testing
