+++ presto-testing/src/main/java/io/prestosql/testing/TestingSessionContext.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.execution;
+package io.prestosql.testing;
 
 import com.google.common.collect.ImmutableMap;
 import io.prestosql.Session;
