+++ presto-tests/src/test/java/io/prestosql/execution/TestQueryRunnerUtil.java
@@ -22,6 +22,7 @@ import io.prestosql.server.BasicQueryInfo;
 import io.prestosql.server.protocol.Slug;
 import io.prestosql.spi.QueryId;
 import io.prestosql.testing.DistributedQueryRunner;
+import io.prestosql.testing.TestingSessionContext;
 
 import java.util.Map;
 import java.util.Set;
