+++ presto-tests/src/test/java/io/prestosql/tests/TestMetadataManager.java
@@ -17,7 +17,6 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import io.prestosql.connector.MockConnectorFactory;
 import io.prestosql.dispatcher.DispatchManager;
-import io.prestosql.execution.TestingSessionContext;
 import io.prestosql.metadata.MetadataManager;
 import io.prestosql.server.BasicQueryInfo;
 import io.prestosql.server.protocol.Slug;
@@ -25,6 +24,7 @@ import io.prestosql.spi.Plugin;
 import io.prestosql.spi.QueryId;
 import io.prestosql.spi.connector.ConnectorFactory;
 import io.prestosql.testing.DistributedQueryRunner;
+import io.prestosql.testing.TestingSessionContext;
 import io.prestosql.tests.tpch.TpchQueryRunnerBuilder;
 import io.prestosql.transaction.TransactionBuilder;
 import org.intellij.lang.annotations.Language;
